backend/dstilh/requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
fasttext==0.9.3
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
pybind11==2.13.6
Pygments==2.19.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
